safari-600.7-branch/LayoutTests/crypto/subtle/hmac-postMessage.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<script>

description("Test sending crypto keys via postMessage.");

jsTestIsAsync = true;

window.addEventListener("message", function(evt) {
    key = evt.data;
    shouldBe("key.type", "'secret'");
    shouldBe("key.extractable", "true");
    shouldBe("key.algorithm.name", "'HMAC'");
    shouldBe("key.algorithm.length", "16");
    shouldBe("key.usages", '["decrypt", "encrypt", "sign", "verify"]');

    finishJSTest();
}, false);

importTestKeys().then(function(keys) {
    postMessage(keys.hmacSha1, "*");
});
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
